requirements.txt
streamlit
streamlit_extras
pandas
numpy
matplotlib
scipy
plotly_express
